[DEV-4963] move es_minimal_sanitize
